// licentainangular.client/src/app/models/NewDTOs.ts
export interface CategoryDTO {
  denumireCategorie: string;
  descriereCategorie: string;
}
// ...
export interface ProdusDTO {
  ProdusId: number;
  Nume: string;
  Descriere?: string;
  EsteSprayText: string;
  Imagine: string | null;
  Pret: number;
  Valabil: boolean;
  Categorie: string;
  Vendor: string;
  IdUser?: number; // <-- adăugat
}